Create a different script for move logging
